pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = ""
authors = [
    {name = "Saemi",email = "cchocolatec@naver.com"}
]
license = {text = "MIT"}
readme = "README.rst"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.2,<6.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-split-settings (>=1.3.2,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
